Move texture format info to its own file
This stuff completely takes over the capability_info file so split it
out into its own.
