+++ src/webgpu/api/operation/command_buffer/image_copy.spec.ts
@@ -41,18 +41,20 @@ TODO: Fix this test for the various skipped formats [2]:
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { assert, memcpy, TypedArrayBufferView, unreachable } from '../../../../common/util/util.js';
 import {
-  kTextureFormatInfo,
-  ColorTextureFormat,
-  kDepthStencilFormats,
   kMinDynamicBufferOffsetAlignment,
   kBufferSizeAlignment,
-  DepthStencilFormat,
-  depthStencilBufferTextureCopySupported,
-  depthStencilFormatAspectSize,
   kTextureDimensions,
-  textureDimensionAndFormatCompatible,
-  kColorTextureFormats,
 } from '../../../capability_info.js';
+import {
+  kTextureFormatInfo,
+  kDepthStencilFormats,
+  kColorTextureFormats,
+  depthStencilBufferTextureCopySupported,
+  textureDimensionAndFormatCompatible,
+  depthStencilFormatAspectSize,
+  DepthStencilFormat,
+  ColorTextureFormat,
+} from '../../../format_info.js';
 import { GPUTest } from '../../../gpu_test.js';
 import { makeBufferWithContents } from '../../../util/buffer.js';
 import { align } from '../../../util/math.js';
