+++ src/webgpu/api/operation/command_buffer/queries/occlusionQuery.spec.ts
@@ -26,7 +26,8 @@ import {
   range,
   unreachable,
 } from '../../../../../common/util/util.js';
-import { kMaxQueryCount, DepthStencilFormat } from '../../../../capability_info.js';
+import { kMaxQueryCount } from '../../../../capability_info.js';
+import { DepthStencilFormat } from '../../../../format_info.js';
 import { GPUTest } from '../../../../gpu_test.js';
 
 const kRequiredQueryBufferOffsetAlignment = 256;
