+++ src/webgpu/api/operation/memory_sync/operation_context_helper.ts
@@ -1,5 +1,5 @@
 import { assert, unreachable } from '../../../../common/util/util.js';
-import { EncodableTextureFormat } from '../../../capability_info.js';
+import { EncodableTextureFormat } from '../../../format_info.js';
 import { GPUTest } from '../../../gpu_test.js';
 
 /**
