+++ src/webgpu/api/operation/render_pass/clear_value.spec.ts
@@ -8,7 +8,7 @@ import {
   kTextureFormatInfo,
   kDepthStencilFormats,
   depthStencilFormatAspectSize,
-} from '../../../capability_info.js';
+} from '../../../format_info.js';
 import { GPUTest } from '../../../gpu_test.js';
 
 export const g = makeTestGroup(GPUTest);
