+++ src/webgpu/api/operation/render_pipeline/culling_tests.spec.ts
@@ -5,7 +5,7 @@ Test all culling combinations of GPUFrontFace and GPUCullMode show the correct o
 `;
 
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
-import { kTextureFormatInfo, SizedTextureFormat } from '../../../capability_info.js';
+import { kTextureFormatInfo, SizedTextureFormat } from '../../../format_info.js';
 import { GPUTest, TextureTestMixin } from '../../../gpu_test.js';
 
 function faceIsCulled(face: 'cw' | 'ccw', frontFace: GPUFrontFace, cullMode: GPUCullMode): boolean {
