+++ src/webgpu/api/operation/rendering/depth_bias.spec.ts
@@ -6,10 +6,10 @@ Tests render results with different depth bias values like 'positive', 'negative
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { unreachable } from '../../../../common/util/util.js';
 import {
+  kTextureFormatInfo,
   DepthStencilFormat,
   EncodableTextureFormat,
-  kTextureFormatInfo,
-} from '../../../capability_info.js';
+} from '../../../format_info.js';
 import { GPUTest, TextureTestMixin } from '../../../gpu_test.js';
 import { TexelView } from '../../../util/texture/texel_view.js';
