+++ src/webgpu/api/operation/texture_view/format_reinterpretation.spec.ts
@@ -4,11 +4,11 @@ Test texture views can reinterpret the format of the original texture.
 
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import {
-  EncodableTextureFormat,
   kRenderableColorTextureFormats,
   kRegularTextureFormats,
   viewCompatible,
-} from '../../../capability_info.js';
+  EncodableTextureFormat,
+} from '../../../format_info.js';
 import { GPUTest, TextureTestMixin } from '../../../gpu_test.js';
 import { TexelView } from '../../../util/texture/texel_view.js';
