+++ src/webgpu/api/validation/capability_checks/features/texture_formats.spec.ts
@@ -4,7 +4,7 @@ Tests for capability checking for features enabling optional texture formats.
 
 import { makeTestGroup } from '../../../../../common/framework/test_group.js';
 import { assert } from '../../../../../common/util/util.js';
-import { kAllTextureFormats, kTextureFormatInfo } from '../../../../capability_info.js';
+import { kAllTextureFormats, kTextureFormatInfo } from '../../../../format_info.js';
 import { kAllCanvasTypes, createCanvas } from '../../../../util/create_elements.js';
 import { ValidationTest } from '../../validation_test.js';
