+++ src/webgpu/api/validation/capability_checks/limits/limit_utils.ts
@@ -2,7 +2,8 @@ import { kUnitCaseParamsBuilder } from '../../../../../common/framework/params_b
 import { makeTestGroup } from '../../../../../common/framework/test_group.js';
 import { getGPU } from '../../../../../common/util/navigator_gpu.js';
 import { assert, range, reorder, ReorderOrder } from '../../../../../common/util/util.js';
-import { kLimitInfo, kTextureFormatInfo } from '../../../../capability_info.js';
+import { kLimitInfo } from '../../../../capability_info.js';
+import { kTextureFormatInfo } from '../../../../format_info.js';
 import { GPUTestBase } from '../../../../gpu_test.js';
 import { align } from '../../../../util/math.js';
