+++ src/webgpu/api/validation/capability_checks/limits/maxColorAttachmentBytesPerSample.spec.ts
@@ -1,5 +1,6 @@
 import { assert } from '../../../../../common/util/util.js';
-import { kTextureSampleCounts, kTextureFormatInfo } from '../../../../capability_info.js';
+import { kTextureSampleCounts } from '../../../../capability_info.js';
+import { kTextureFormatInfo } from '../../../../format_info.js';
 import { align } from '../../../../util/math.js';
 
 import {
