+++ src/webgpu/api/validation/image_copy/buffer_texture_copies.spec.ts
@@ -5,14 +5,13 @@ the general image_copy tests, or by destroyed,*.
 
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { assert, unreachable } from '../../../../common/util/util.js';
+import { kBufferUsages, kTextureUsages } from '../../../capability_info.js';
+import { GPUConst } from '../../../constants.js';
 import {
   kDepthStencilFormats,
-  kBufferUsages,
-  kTextureUsages,
   depthStencilBufferTextureCopySupported,
   depthStencilFormatAspectSize,
-} from '../../../capability_info.js';
-import { GPUConst } from '../../../constants.js';
+} from '../../../format_info.js';
 import { align } from '../../../util/math.js';
 import { kBufferCopyAlignment, kBytesPerRowAlignment } from '../../../util/texture/layout.js';
 import { ValidationTest } from '../validation_test.js';
