+++ src/webgpu/api/validation/image_copy/image_copy.ts
@@ -1,9 +1,9 @@
 import {
-  kTextureFormatInfo,
-  SizedTextureFormat,
-  DepthStencilFormat,
   depthStencilFormatCopyableAspects,
-} from '../../../capability_info.js';
+  DepthStencilFormat,
+  SizedTextureFormat,
+  kTextureFormatInfo,
+} from '../../../format_info.js';
 import { align } from '../../../util/math.js';
 import { ImageCopyType } from '../../../util/texture/layout.js';
 import { ValidationTest } from '../validation_test.js';
