+++ src/webgpu/api/validation/image_copy/layout_related.spec.ts
@@ -4,12 +4,12 @@ TODO check if the tests need to be updated to support aspects of depth-stencil t
 
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { assert } from '../../../../common/util/util.js';
+import { kTextureDimensions } from '../../../capability_info.js';
 import {
   kTextureFormatInfo,
   kSizedTextureFormats,
   textureDimensionAndFormatCompatible,
-  kTextureDimensions,
-} from '../../../capability_info.js';
+} from '../../../format_info.js';
 import { align } from '../../../util/math.js';
 import {
   bytesInACompleteRow,
