+++ src/webgpu/api/validation/queue/copyToTexture/CopyExternalImageToTexture.spec.ts
@@ -11,12 +11,12 @@ import {
 } from '../../../../../common/framework/resources.js';
 import { makeTestGroup } from '../../../../../common/framework/test_group.js';
 import { raceWithRejectOnTimeout, unreachable, assert } from '../../../../../common/util/util.js';
+import { kTextureUsages } from '../../../../capability_info.js';
 import {
   kTextureFormatInfo,
   kTextureFormats,
-  kTextureUsages,
   kValidTextureFormatsForCopyE2T,
-} from '../../../../capability_info.js';
+} from '../../../../format_info.js';
 import { kResourceStates } from '../../../../gpu_test.js';
 import {
   CanvasType,
