+++ src/webgpu/api/validation/state/device_lost/destroy.spec.ts
@@ -13,17 +13,19 @@ import {
   kBufferUsageInfo,
   kBufferUsageCopy,
   kBufferUsageCopyInfo,
-  kCompressedTextureFormats,
   kQueryTypes,
   kTextureUsageType,
   kTextureUsageTypeInfo,
   kTextureUsageCopy,
   kTextureUsageCopyInfo,
+  kShaderStageKeys,
+} from '../../../../capability_info.js';
+import {
+  kCompressedTextureFormats,
   kRegularTextureFormats,
   kRenderableColorTextureFormats,
-  kShaderStageKeys,
   kTextureFormatInfo,
-} from '../../../../capability_info.js';
+} from '../../../../format_info.js';
 import { CommandBufferMaker, EncoderType } from '../../../../util/command_buffer_maker.js';
 import {
   createCanvas,
