+++ src/webgpu/web_platform/canvas/configure.spec.ts
@@ -8,16 +8,15 @@ TODO:
 
 import { makeTestGroup } from '../../../common/framework/test_group.js';
 import { assert } from '../../../common/util/util.js';
+import { kCanvasTextureFormats, kTextureUsages } from '../../capability_info.js';
+import { GPUConst } from '../../constants.js';
 import {
   kAllTextureFormats,
-  kCanvasTextureFormats,
-  kTextureUsages,
-  filterFormatsByFeature,
   kFeaturesForFormats,
   kTextureFormats,
+  filterFormatsByFeature,
   viewCompatible,
-} from '../../capability_info.js';
-import { GPUConst } from '../../constants.js';
+} from '../../format_info.js';
 import { GPUTest } from '../../gpu_test.js';
 import { kAllCanvasTypes, createCanvas } from '../../util/create_elements.js';
