+++ src/webgpu/web_platform/copyToTexture/ImageBitmap.spec.ts
@@ -15,7 +15,7 @@ import {
   kTextureFormatInfo,
   kValidTextureFormatsForCopyE2T,
   EncodableTextureFormat,
-} from '../../capability_info.js';
+} from '../../format_info.js';
 import { CopyToTextureUtils, kCopySubrectInfo } from '../../util/copy_to_texture.js';
 import { PerTexelComponent } from '../../util/texture/texel_data.js';
 import { TexelView } from '../../util/texture/texel_view.js';
