+++ src/webgpu/web_platform/copyToTexture/canvas.spec.ts
@@ -4,12 +4,12 @@ copyToTexture with HTMLCanvasElement and OffscreenCanvas sources.
 
 import { makeTestGroup } from '../../../common/framework/test_group.js';
 import { skipTestCase } from '../../../common/util/util.js';
+import { kCanvasAlphaModes } from '../../capability_info.js';
 import {
-  kCanvasAlphaModes,
   kTextureFormatInfo,
   kValidTextureFormatsForCopyE2T,
   RegularTextureFormat,
-} from '../../capability_info.js';
+} from '../../format_info.js';
 import { CopyToTextureUtils } from '../../util/copy_to_texture.js';
 import { CanvasType, kAllCanvasTypes, createCanvas } from '../../util/create_elements.js';
 import { TexelCompareOptions } from '../../util/texture/texture_ok.js';
