+++ src/webgpu/web_platform/reftests/canvas_complex.html.ts
@@ -1,5 +1,5 @@
 import { assert, unreachable } from '../../../common/util/util.js';
-import { ColorTextureFormat, kTextureFormatInfo } from '../../capability_info.js';
+import { kTextureFormatInfo, ColorTextureFormat } from '../../format_info.js';
 import { gammaDecompress, float32ToFloat16Bits } from '../../util/conversion.js';
 import { align } from '../../util/math.js';
